// src/app/models/customer.ts
export interface Customer {
  names?: string;
  address?: string;
  numHistory?: number;
  phone?: string;
  dni?: string;
  country?: string;
  email?: string;
  password?: string;
  linkWathsapp?: string;
  version?: string;
  verificado?: boolean;
  activeVirtual?: boolean;
  qualification?: number;
  age?: number;
  consultant: boolean;
  id?: string;
  uid?: string;
}